feat: make emacs readline keybindings configurable
